pyproject.toml
[tool.poetry]
name = "design_perma"
version = "0.1.0"
description = ""
authors = ["Vincent Lefoulon <vincent.lefoulon@free.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"
django-filter = "^2.4.0"
shapely = "^1.7.1"
jsonfield = "^3.1.0"
django-categories = "^1.8"
djangorestframework-recursive = "^0.1.2"
django-inline-actions = "^2.3.0"
django-json-widget = "^1.0.1"
jsonschema = "^3.2.0"
django-jsonschema-form = "^1.0.3"
django-polymorphic = "^3.0.0"
django-environ = "^0.4.5"
requests = "^2.25.0"
django-tagging = "^0.5.0"
django-colorfield = "^0.3.2"
pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pre-commit = "^2.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
